[Refactor] 커스텀 KeyboardAnimation를 Granite KeyboardAboveView로 대체 (#3)
